Make attack time basically match the display
The attack time, with its 0.9 peak and coeffients set to
a 1.0 overshoot. didn't match the display time. This fixes it by
adjusting the attack time in the coefficient appropriately.

And also stable under modulation

Addresses #116

Still need to check release and decay same way
And still need to move to basic blocks metadat