Rename TriStateButton to MultiStateButton, now it works! (#198)
* Rename TriStateButton to MultiStateButton, now it works!

It made sense to make it a generic multi-state button instead of fixing it to just three states. We might not need it for OB-Xf but it's good to have around. Closes #122!

The issue was not calling setValue() in the widget, and range in PMD not being quantized.

Also implemented voice status LEDs on the GUI.

Fixed red noise member variable not being initialized.

When assigning things to componentMap, don't repeat the name string we've searched for already at the beginning of the block - just use the `name` variable!