Make the PROJECT_VERSION advance with date (#235)
The PROJECT_VERSION was 0.8 but this changes it to
0.(year-2025+5).(day-of-year+1) meaning we get a new
version each day.

This is particularly useful for macos users where the
au hosting service uses version number advances to trigger
rescans and the like properly.

But also is just a useful practice from SC borrowed here.

Obvioiusly when we ship 1.0 go to manual version numbers again