Dont apply param changes which aren't changes in audio thread (#242)
This changes two things.

First, the ParamManger only does a callback if the value has changed
on a param set unless forced. The attachment setup when the editor
opens forces all params to send their value which triggers this,
and we had reports of a sound when opening editor which matches
the 'set-same-is-not-idempotent'. Rather than chase each setter,
just block here.

Second, it cleans up some of the debug string math which was very
allocaty on the audio thread.