Force execute SetHQ() when calling Motherboard::setSampleRate()
This fixes the issue where renders sounded an octave higher if HQ button was enabled.
The cause was a SR mismatch because JUCE's prepareToPlay sent us host's SR (say 48k), this propagated throughout the synth, but HQ stuff was not being executed because Motherboard::SetHQ() was guarded by an early return if incoming value was the same as the `oversample` member.

Added a force argument to Motherboard::SetHQ() which is then only called from Motherboard::setSampleRate().

Fixes #269!