Restore automation->ui channel
This change upgrades the interaction between automation and the UI
in a couple of ways

First it makes the param manager listener a set for multiple purposes
so we can have both an engine and a UI listener for a given param

Second it makes the param listener thread safe in the UI side - we
are lucky that code was never called!

Third, for knobs, it bundles begin/end edits around the entire drag
not around each update

Upshot is: Drag the knob in bitwig and the screen in obxf updates
as well as the engine.

CLoses #289