Fix a problem with current program and bank loads (#332)
A bank load would send a daw notification loop for *every* program
not just current. This caused all sorts of pain, including defacto
setting the current program to 255 values no matter what.

So instead just set the parameters in the engine and host to the
parameters for the current program when loading an entire block.

Worked on startup since that did a setCurrentPrgam 0 before process
started

Closes #290

And with this we have the bank mechanics basically done.

So Closes #313