Be more parsimonious in our use of updateHostDisplay
updateHostDisplay by default udpates everything - latency
param names and values programs and so on. We were calling
it willy nilly with the default. This over-stressed all daws
but in Logic it meant the setCurrentProgram order of operations
would cause a rescan of params without them having been host updated
and that would reset the entire thing wrongly. Bad bad.

So change it so that doesn't happen and low logic works great!

Addresses #333