Fix some initial zoom problems
Initial zoom of non-100% didn't really work and I had the
odd delayed resized call in the idle loop to work around it.
Well I found out the problem. Basically

1. Want to force a call to scaleFactorChanged and
2. Label was screwed up generlally

Fix 1 by adding a HasScaleFactor interface we can cast to
to force post resize in constructor

Fix 2 by moving Label for a Drawable to a Component. There
was no need for it to ever implement the Drawable API.