Restore the resize on idle
Something is making VST3 on windows high def lose a resize
message. I have a theory that the queue gets flooded while
loading assets and one gets lost, or a resize gets called
before a size is set, or some such. But I can't pin it down.
This restores the resize on idle hammer.