Replace RMB on LFO target buttons with Ctrl/Cmd+click (#442)
This will allow us to use context menu on buttons down the line if we wanted it (for consistency with knobs, we have the same in Surge XT).

Also, make it so that Ctrl/Cmd+clicking the 16 patch switching buttons will behave inverse to the state of the Group button.