Logic load sequence and FIFO slop fix (#488)
Our use of the FIFO assumed it was 'always drained' in a couple of
spots and that meant in the logic pro (correct) load sequence that
assumption failed us. So we would apply partial or incomplete states
and have all sorts of other nasties.

This does two things

1. On a force load, still process the fifo if outstanding for callbacks
   so when prepareToPlay interacts with a load still going on we dont
   drop the load and
2. In the UI side loadCurrentProgram which updates all the params
   if there's still unprocessed fifo kicking around, defer the loadCurrent
   until that queue is all clear.

Should address #477